Move allowed value classes
- Moved allowed values classes
